requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2020.4.5.1
gast==0.2.2
google-pasta==0.1.8
gpytorch==1.0.1
grpcio==1.25.0
h5py==2.10.0
importlib-metadata==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.17.4
Pillow==6.2.1
protobuf==3.11.1
scipy==1.3.3
six==1.13.0
tensorboard==1.15.0
tensorboardX==2.0
tensorflow==1.15.0
tensorflow-compression==1.3
tensorflow-estimator==1.15.1
tensorflow-probability==0.7.0
